Rename tabel to table
